frontend/requirements.txt
streamlit
requests
PyYAML
pyjwt
